Move parser util to a separate file
