+++ src/components/Home.js
@@ -4,6 +4,7 @@ import { useHistory } from 'react-router';
 import Settings from './Settings';
 import DataSource from './DataSource';
 import { DataSources } from '../constants';
+import { parseCustomWords } from '../utils/words';
 import { useWordsState } from '../contexts/words';
 import { useSettingsState } from '../contexts/settings';
 
@@ -47,56 +48,14 @@ function Home() {
     <div className="flex-wrap">
       <DataSource {...{ words, settings, dataSource, customWords, setDataSource, setCustomWords }} />
       <Settings settings={settings} setSettings={setSettings} />
-      <button className="float-right py-2 px-4 font-semibold rounded-lg shadow-md text-white bg-green-500 hover:bg-green-700" onClick={onSubmit}>
+      <button
+        className="float-right py-2 px-4 font-semibold rounded-lg shadow-md text-white bg-green-500 hover:bg-green-700"
+        onClick={onSubmit}
+      >
         Submit
       </button>
     </div>
   );
-}
-
-/**
- * Parses and validates the passed custom words and throws appropriate errors.
- *
- * @param {string} customWords
- * @param {boolean} shouldShowMeaning
- * @returns {object}
- */
-function parseCustomWords(customWords, shouldShowMeaning) {
-  let parsedCustomWords;
-
-  try {
-    parsedCustomWords = JSON.parse(customWords);
-  } catch (e) {
-    parsedCustomWords = customWords.split(',').reduce((acc, word) => {
-      const trimmedWord = word.trim();
-
-      if (trimmedWord) {
-        acc.push({ word: trimmedWord });
-      }
-
-      return acc;
-    }, []);
-  }
-
-  if (!Array.isArray(parsedCustomWords)) {
-    throw new Error('Custom words needs to be an array.');
-  }
-
-  for (let { word, meaning } of parsedCustomWords) {
-    if (!word) {
-      throw new Error(
-        `Every element in passed data must have a "word" ${shouldShowMeaning ? 'as well as a "meaning"' : ''} property.`
-      );
-    }
-
-    if (shouldShowMeaning && !meaning) {
-      throw new Error(
-        `Every element in passed data must have a "meaning" property.\n\nTurn off "Display meaning of current word" setting to continue without "meaning" property.`
-      );
-    }
-  }
-
-  return parsedCustomWords;
 }
 
 export default Home;
